refactor(telemetry): move uriFilter out of recording/ to decouple EQ
