+++ extension/src/extension/services/telemetry/recording/snapshots/snapshotManager.ts
@@ -4,7 +4,7 @@ import { LogCategory, logger } from '@extension/services/loggingService';
 import type { RecorderLifecycleState } from '@extension/services/telemetry/recording/lifecycleController';
 import type { RecordingStorageWriter } from '@extension/services/telemetry/recording/storageWriter';
 import type { FileSnapshotErrorEvent, RecordedEvent } from '@extension/services/telemetry/recording/types';
-import { shouldRecordUri } from '@extension/services/telemetry/recording/uriFilter';
+import { shouldRecordUri } from '@extension/services/telemetry/uriFilter';
 
 interface SnapshotManagerDeps {
     state: RecorderLifecycleState;
