+++ extension/src/extension/services/telemetry/uriFilter.ts
